+++ topiary-core/src/atom_collection.rs
@@ -796,7 +796,7 @@ impl AtomCollection {
                 {
                     vec.push(atom);
                 } else {
-                    log::warn!("Found scoped softline {:?} outside of its scope", atom);
+                    log::warn!("Found scoped softline {atom:?} outside of its scope");
                     force_apply_modifications = true;
                 }
             // Register the ScopedConditional in the correct scope
@@ -806,7 +806,7 @@ impl AtomCollection {
                 {
                     vec.push(atom);
                 } else {
-                    log::warn!("Found scoped conditional {:?} outside of its scope", atom);
+                    log::warn!("Found scoped conditional {atom:?} outside of its scope");
                     force_apply_modifications = true;
                 }
             }
@@ -816,7 +816,7 @@ impl AtomCollection {
             .filter_map(|(scope_id, vec)| if vec.is_empty() { None } else { Some(scope_id) })
             .collect();
         if !still_opened.is_empty() {
-            log::warn!("Some scopes have been left opened: {:?}", still_opened);
+            log::warn!("Some scopes have been left opened: {still_opened:?}");
             force_apply_modifications = true;
         }
         still_opened = opened_measuring_scopes
@@ -824,10 +824,7 @@ impl AtomCollection {
             .filter_map(|(scope_id, vec)| if vec.is_empty() { None } else { Some(scope_id) })
             .collect();
         if !still_opened.is_empty() {
-            log::warn!(
-                "Some measuring scopes have been left opened: {:?}",
-                still_opened
-            );
+            log::warn!("Some measuring scopes have been left opened: {still_opened:?}");
             force_apply_modifications = true;
         }
 
@@ -850,10 +847,7 @@ impl AtomCollection {
                     if let Some(replacement) = modifications.remove(id) {
                         *atom = replacement;
                     } else {
-                        log::warn!(
-                            "Found scoped softline {:?}, but was unable to replace it.",
-                            atom
-                        );
+                        log::warn!("Found scoped softline {atom:?}, but was unable to replace it.");
                         *atom = Atom::Empty;
                     }
                 } else if let Atom::ScopedConditional { id, .. } = atom {
@@ -861,8 +855,7 @@ impl AtomCollection {
                         *atom = replacement;
                     } else {
                         log::warn!(
-                            "Found scoped conditional {:?}, but was unable to replace it.",
-                            atom
+                            "Found scoped conditional {atom:?}, but was unable to replace it."
                         );
                         *atom = Atom::Empty;
                     }